requirements.txt
asgiref==3.5.2
Django==4.1.1
sqlparse==0.4.3
gunicorn
djangorestframework
django-environ==0.8.1
psycopg2